arreglo cono vision

+++ server/game_config.cpp
@@ -12,7 +12,7 @@ GameConfig::GameConfig(int player_health, int number_of_rounds, int starting_mon
                        GunConfig&& m3_config, int defuse_time, int time_to_plant, int bomb_dmg,
                        int round_winner_money, int round_loser_money, int buy_time, int bomb_time,
                        int after_round_time, int money_per_kill, int tiles_per_movement,
-                       int game_rate, MapConfig&& map_config, float cone_angle, int cone_opacity):
+                       int game_rate, MapConfig&& map_config, int cone_angle, int cone_opacity):
         player_health(player_health),
         number_of_rounds(number_of_rounds),
         starting_money(starting_money),
