+++ server/game_config.h
@@ -33,7 +33,7 @@ public:
     const int tiles_per_movement;
     const int game_rate;
     const MapConfig map_config;
-    const float cone_angle;
+    const int cone_angle;
     const int cone_opacity;
     GameConfig(int player_health, int number_of_rounds, int starting_money, int ct_amount,
                int tt_amount, int ammo_price, GunConfig&& knife_config, GunConfig&& glock_config,
@@ -41,7 +41,7 @@ public:
                int defuse_time, int time_to_plant, int bomb_dmg, int round_winner_money,
                int round_loser_money, int buy_time, int bomb_time, int after_round_time,
                int money_per_kill, int tiles_per_movement, int game_rate, MapConfig&& map_config,
-               float cone_angle, int cone_opacity);
+               int cone_angle, int cone_opacity);
 };
